+++ solc/CommandLineParser.cpp
@@ -896,6 +896,9 @@ void CommandLineParser::parseArgs(int _argc, char const* const* _argv)
 	po::options_description allOptions = optionsDescription();
 	po::positional_options_description filesPositions = positionalOptionsDescription();
 
+	m_options = {};
+	m_args = {};
+
 	// parse the compiler arguments
 	try
 	{
@@ -914,6 +917,11 @@ void CommandLineParser::parseArgs(int _argc, char const* const* _argv)
 
 void CommandLineParser::processArgs()
 {
+	if (m_args.count(g_strNoColor) > 0)
+		m_options.formatting.coloredOutput = false;
+	else if (m_args.count(g_strColor) > 0)
+		m_options.formatting.coloredOutput = true;
+
 	checkMutuallyExclusive({
 		g_strHelp,
 		g_strLicense,
@@ -1027,11 +1035,6 @@ void CommandLineParser::processArgs()
 				"Option --" + g_strDebugInfo + " is only valid in compiler and assembler modes."
 			);
 	}
-
-	if (m_args.count(g_strColor) > 0)
-		m_options.formatting.coloredOutput = true;
-	else if (m_args.count(g_strNoColor) > 0)
-		m_options.formatting.coloredOutput = false;
 
 	m_options.formatting.withErrorIds = m_args.count(g_strErrorIds);