+++ test/cmdlineTests/~unknown_options/test.sh
@@ -9,7 +9,7 @@ output=$("$SOLC" --allow=test 2>&1)
 failed=$?
 set -e
 
-if [[ $output != "unrecognised option '--allow=test'" ]] || (( failed == 0 ))
+if [[ $output != "Error: unrecognised option '--allow=test'" ]] || (( failed == 0 ))
 then
     fail "Incorrect response to unknown options: $output"
 fi